# openapi.yml
openapi: 3.0.3
info:
  title: Luga Store API
  version: 1.0.0
  description: API documentation for the Luga Store backend, listing all customer, shop manager/owner, and administrator endpoints.
security:
  - CookieAuth: []
paths:
...
  /invitations/accept-owner:
    post:
      tags: [Invitations]
      summary: Accept a shop owner invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptShopMemberInvitationRequest'
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /invitations/accept-manager:
    post:
      tags: [Invitations]
      summary: Accept a shop manager invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptShopMemberInvitationRequest'
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /invitations/accept-admin:
    post:
      tags: [Invitations]
      summary: Accept an admin invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptAdminInvitationRequest'
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /invitations/owner-details:
    get:
      tags: [Invitations]
      summary: Get details of a shop owner invitation
      security: []
      parameters:
        - name: shopId
          in: query
          required: true
          schema:
            type: integer
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationDetails'
  /invitations/manager-details:
    get:
      tags: [Invitations]
      summary: Get details of a shop manager invitation
      security: []
      parameters:
        - name: shopId
          in: query
          required: true
          schema:
            type: integer
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationDetails'
  /invitations/admin-details:
    get:
      tags: [Invitations]
      summary: Get details of an admin invitation
      security: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInvitationDetails'

  # --- GLOBAL CATEGORIES (Admin/Owner/Manager) ---
...
  /products:
    get:
      tags: [Products (Global)]
      summary: Get list of products
      security: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: shopId
          in: query
          schema:
            type: integer
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of products

    post:
      tags: [Products (Global)]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200':
          description: Product created successfully
  /products/{id}:
    get:
      tags: [Products (Global)]
      summary: Get details of a single product
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details

    put:
      tags: [Products (Global)]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200':
          description: Product updated successfully

    delete:
      tags: [Products (Global)]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Product deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /products/{id}/sizes:
    patch:
      tags: [Products (Global)]
      summary: Sync product size inventories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductSizesRequest'
      responses:
        '200':
          description: Product sizes updated successfully
  /products/{id}/status:
    patch:
      tags: [Products (Global)]
      summary: Change product active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductActiveStatusRequest'
      responses:
        '200':
          description: Status updated successfully

  # --- SHOP-LEVEL PRODUCTS ---
...
  /shops/{shopId}/products/{id}:
    put:
      tags: [Products (Shop)]
      summary: Update a product in a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200':
          description: Product updated successfully

    delete:
      tags: [Products (Shop)]
      summary: Delete a product in a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Product deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /shops/{shopId}/products/{id}/status:
    patch:
      tags: [Products (Shop)]
      summary: Toggle active status of a product in a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductActiveStatusRequest'
      responses:
        '200':
          description: Product status updated successfully

  # --- CUSTOMER PROFILE ---
  # --- CUSTOMER PROFILE ---
  /profile:
    get:
      tags: [Profile]
      summary: Get the authenticated user's profile with all roles
      responses:
        '200':
          description: User profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    patch:
      tags: [Profile]
      summary: Update the authenticated user's profile details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
      '200':
        description: Profile updated successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /profile/password:
    patch:
      tags: [Profile]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeUserPasswordRequest'
      responses:
      '200':
        description: Password changed successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /profile/shops:
    get:
      tags: [Profile]
      summary: Get shops managed or owned by the current user
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of user owned/managed shops

  # --- CUSTOMER ADDRESSES ---
...
  /media/upload-url:
    post:
      tags: [Media]
      summary: Request presigned S3 URLs to upload files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileS3UploadUrlRequest'
      responses:
        '200':
          description: Upload URLs generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileUploadUrl'
  /media/{filename}:
    delete:
      tags: [Media]
      summary: Delete an uploaded file
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
      '200':
        description: File deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string

  # --- NEWSLETTER ---
...
  /shops:
    get:
      tags: [Shops]
      summary: Get paginated list of shops (Admin only)
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of shops

    post:
      tags: [Shops]
      summary: Create a new shop (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertShopRequest'
      responses:
        '200':
          description: Shop created successfully
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application
...
  schemas:
...
    AcceptShopMemberInvitationRequest:
      type: object
      required: [shopId, email, token, firstName, lastName, password]
      properties:
        shopId:
          type: integer
        email:
          type: string
          format: email
        token:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
    AcceptAdminInvitationRequest:
      type: object
      required: [email, token, firstName, lastName, password]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
    InvitationDetails:
      type: object
      properties:
        shopId:
          type: integer
        shopName:
          type: string
        email:
          type: string
    AdminInvitationDetails:
      type: object
      properties:
        email:
          type: string
...
    UpsertProductRequest:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        gender:
          type: string
          enum: [Men, Women, Unisex]
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
    UpdateProductSizesRequest:
      type: object
      required: [sizes]
      properties:
        sizes:
          type: array
          items:
            $ref: '#/components/schemas/ProductSizeDto'
    ProductSizeDto:
      type: object
      required: [size, stock]
      properties:
        size:
          type: string
          enum: [XS, S, M, L, XL, XXL, XXXL]
        stock:
          type: integer
    UpdateProductActiveStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
    UserDto:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        avatarFileName:
          type: string
          nullable: true
        isEmailConfirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
...
    ChangeUserPasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
...
    UpsertShopRequest:
      type: object
      required: [name, subdomainSlug]
      properties:
        name:
          type: string
        subdomainSlug:
          type: string
        description:
          type: string
        logoFileName:
          type: string
        bannerFileName:
          type: string
...
    FileS3UploadUrlRequest:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileMeta'
    FileMeta:
      type: object
      required: [fileName, contentType]
      properties:
        fileName:
          type: string
        contentType:
          type: string
    FileUploadUrl:
      type: object
      properties:
        fileName:
          type: string
        uploadUrl:
          type: string
...